Movendo algumas funcoes pra utils.py
